requirements.txt
setuptools
pandas
numpy
google-cloud-storage
pyyaml
tensorflow
joblib
comet-ml
dvc